# interp: Specialized, optimized implementations for common interpolations