--- a/benches/interp.rs
+++ b/benches/interp.rs
@@ -72,6 +72,38 @@ pub fn criterion_benchmark(c: &mut Criterion) {
     }
 
     group.finish();
+
+    let mut group = c.benchmark_group("interp process: 49 taps 4 factors 2ch");
+    let mut data_out = vec![0.0f32; 48_000 * 5 * 2 * 4];
+
+    #[cfg(feature = "c-tests")]
+    unsafe {
+        let interp = interp::interp_create_c(
+            black_box(49),
+            black_box(data_out.len() / data.len()) as u32,
+            black_box(2),
+        );
+        group.bench_function("C", |b| {
+            b.iter(|| {
+                interp::interp_process_c(interp, 48_000 * 5, data.as_ptr(), data_out.as_mut_ptr());
+            })
+        });
+        interp::interp_destroy_c(interp);
+    }
+    {
+        let mut interp = interp::Interp::new(
+            black_box(49),
+            black_box(data_out.len() / data.len()),
+            black_box(2),
+        );
+        group.bench_function("Rust", |b| {
+            b.iter(|| {
+                interp.process(&data, &mut data_out);
+            })
+        });
+    }
+
+    group.finish();
 }
 
 criterion_group!(benches, criterion_benchmark);