@sdroege 2020-10-15
Thanks! I'll review this in the next days. Switching the coefficients to `f32` seems ok, we already store everything else as `f32` anyway so the improved precision would only be very transient. I'll check how much effect that has in detail though.

If you have any other questions before I do a proper review, please just ask :)