@sdroege 2020-10-15
Can you add some numbers in the commit message (and PR description) from the benchmarks you did?