@rawler 2020-10-15
> Can you add some numbers in the commit message (and PR description) from the benchmarks you did?

Sure. I chose not to do it initially, since I have a little bit of a problem with benchmarks on my laptop (throttling causes some uncertainty). But here's a best effort attached.