@sdroege 2020-10-15
No problem, I can run it here later and update it with my numbers if they're more stable :)