@sdroege 2020-10-22
Thanks for the update. I'll go through the code another time later as it changed quite a bit since last time