@rawler 2020-10-24
There. I think I've resolved all open questions, and from my side I'm ready to merge?

A showed this PR to a colleague who came up with yet another great optimization to this code, but I'd say that can be a PR on it's own, this PR is already doing more than I want, really :)